Almost there. Movement is not rendering correctly so there's a bug, probably in Add handling.
Oh, now the bug is somewhere between render and layout effect. Solved: the real problem is that ref wasn't actually being set during render because we were using an ad hoc (via unboxing) version of the ref prop instead
of the strongly-typed version in Konva.fs. Not sure yet what the difference was but that was the proximate error.

Currently animation works, but the values are wrong somehow. Text moves more accurately than circle does.